backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
drf-spectacular==0.27.0
psycopg2-binary>=2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
google-generativeai==0.8.3
Pillow>=10.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
